+++ pandas/conftest.py
@@ -931,13 +931,3 @@ def non_mapping_dict_subclass():
             return self._data.__len__()
 
     return TestNonDictMapping
-
-
-def async_mark():
-    try:
-        import_optional_dependency("pytest_asyncio")
-        async_mark = pytest.mark.asyncio
-    except ImportError:
-        async_mark = pytest.mark.skip(reason="Missing dependency pytest-asyncio")
-
-    return async_mark
